+++ src/django_browser_reload/templatetags/django_browser_reload.py
@@ -1,8 +1,14 @@
 from __future__ import annotations
 
-from django import template
+from django.template import Context, Library
 
-from django_browser_reload.jinja import django_browser_reload_script
+from django_browser_reload.jinja import (
+    django_browser_reload_script as base_django_browser_reload_script,
+)
 
-register = template.Library()
-register.simple_tag(django_browser_reload_script)
+register = Library()
+
+
+@register.simple_tag(takes_context=True)
+def django_browser_reload_script(context: Context) -> str:
+    return base_django_browser_reload_script(nonce=context.get("csp_nonce"))