+++ tests/templatetags/test_django_browser_reload.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+import secrets
+
 from django.template import Context, Template
 from django.test import SimpleTestCase, override_settings
 
@@ -23,4 +25,20 @@ class DjangoBrowserReloadScriptTests(SimpleTestCase):
             + ' data-worker-script-path="/static/django-browser-reload/'
             + 'reload-worker.js"'
             + ' data-events-path="/__reload__/events/" defer></script>'
+        )
+
+    def test_debug_nonce(self):
+        nonce = secrets.token_urlsafe(16)
+
+        with override_settings(DEBUG=True):
+            result = Template(
+                "{% load django_browser_reload %}{% django_browser_reload_script %}"
+            ).render(Context({"csp_nonce": nonce}))
+
+        assert result == (
+            '<script src="/static/django-browser-reload/reload-listener.js"'
+            + ' data-worker-script-path="/static/django-browser-reload/'
+            + 'reload-worker.js"'
+            + f' data-events-path="/__reload__/events/" defer nonce="{nonce}">'
+            + "</script>"
         )